requirements.txt
streamlit==1.32.2
matplotlib==3.8.3
pandas==2.2.1
numpy==1.26.4
scipy
streamlit-option-menu
pyarrow
anthropic
tavily-python
tweepy
google-generativeai
python-dotenv
